+++ conanfile.py
@@ -1,9 +1,11 @@
 from conans import ConanFile, CMake
+from conans.tools import load
+import re
+import os
 
 
 class bertrandConan(ConanFile):
     name = "bertrand"
-    version = "0.0.12"
     license = "LGPLv3"
     author = "Dominik Berner <dominik.berner+bertrand-conan@gmail.com"
     url = "https://github.com/bernedom/bertrand"
@@ -20,6 +22,13 @@ class bertrandConan(ConanFile):
         # Add additional settings with cmake.definitions["SOME_DEFINITION"] = True
         cmake.configure()
         return cmake
+
+    def set_version(self):
+        cmake = load(os.path.join(self.recipe_folder, "CMakeLists.txt"))
+
+        version = re.search(
+            r"(?:[ \t]*)(?:VERSION\s+?)(\d+\.\d+\.\d+)", cmake).group(1)
+        self.version = version
 
     def build(self):
         cmake = self._configure_cmake()