+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.0.12 
 
 *   Contract type "Require", "Ensure" or "Invariant" is included in message
+*   conanfile.py pulls version out of CMakeLists
+*   Linting conanfile.py, cmake object is only generated once
 
 ## 0.0.11